requirements.txt
discord.py[voice]==1.6.0
PyOgg==0.6.14a1